backEnd/requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.46.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
datasets==4.0.0
dill==0.3.8
distro==1.9.0
evaluate==0.4.5
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
groq==0.30.0
gTTS==2.5.4
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.1
peft==0.16.0
playsound==1.3.0
propcache==0.3.2
psutil==7.0.0
pyarrow==21.0.0
PyAudio 
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.31
requests==2.32.4
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
speechbrain==1.0.3
starlette==0.47.2
sympy==1.14.0
tokenizers==0.21.4
torch==2.7.1
torchaudio==2.7.1
torchcodec==0.5
tqdm==4.67.1
transformers==4.54.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.5.0
yarl==1.20.1
